move shared github models to shared
